rename utils file and move data collector file
