+++ tests/backtest_data_collector_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from afang.backtest_data_collector import (
+from afang.database.backtest_data_collector import (
     collect_all,
     fetch_initial_data,
     fetch_most_recent_data,
@@ -32,13 +32,13 @@ def test_fetch_initial_data(
 ) -> None:
     # mock the get_historical_data function
     mocked_get_historical_data = mocker.patch(
-        "afang.backtest_data_collector.IsExchange.get_historical_data"
+        "afang.database.backtest_data_collector.IsExchange.get_historical_data"
     )
     mocked_get_historical_data.return_value = historical_data
 
     # mock the write_data function
     mocked_write_data = mocker.patch(
-        "afang.backtest_data_collector.OHLCVDatabase.write_data"
+        "afang.database.backtest_data_collector.OHLCVDatabase.write_data"
     )
 
     test_ohlcv_db = OHLCVDatabase(ohlcv_root_db_dir, "test_exchange", "test_symbol")
@@ -63,13 +63,13 @@ def test_fetch_most_recent_data(
 ) -> None:
     # mock the get_historical_data function
     mocked_get_historical_data = mocker.patch(
-        "afang.backtest_data_collector.IsExchange.get_historical_data"
+        "afang.database.backtest_data_collector.IsExchange.get_historical_data"
     )
     mocked_get_historical_data.return_value = historical_data
 
     # mock the write_data function
     mocked_write_data = mocker.patch(
-        "afang.backtest_data_collector.OHLCVDatabase.write_data"
+        "afang.database.backtest_data_collector.OHLCVDatabase.write_data"
     )
 
     test_ohlcv_db = OHLCVDatabase(ohlcv_root_db_dir, "test_exchange", "test_symbol")
@@ -96,13 +96,13 @@ def test_fetch_older_data(
 ) -> None:
     # mock the get_historical_data function
     mocked_get_historical_data = mocker.patch(
-        "afang.backtest_data_collector.IsExchange.get_historical_data"
+        "afang.database.backtest_data_collector.IsExchange.get_historical_data"
     )
     mocked_get_historical_data.return_value = historical_data
 
     # mock the write_data function
     mocked_write_data = mocker.patch(
-        "afang.backtest_data_collector.OHLCVDatabase.write_data"
+        "afang.database.backtest_data_collector.OHLCVDatabase.write_data"
     )
 
     test_ohlcv_db = OHLCVDatabase(ohlcv_root_db_dir, "test_exchange", "test_symbol")
@@ -121,31 +121,31 @@ def test_collect_all(mocker, dummy_is_exchange, ohlcv_root_db_dir):
 
     # mock the get_min_max_timestamp function
     mocked_get_min_max_timestamp = mocker.patch(
-        "afang.backtest_data_collector.OHLCVDatabase.get_min_max_timestamp"
+        "afang.database.backtest_data_collector.OHLCVDatabase.get_min_max_timestamp"
     )
     mocked_get_min_max_timestamp.return_value = (None, None)
 
     # mock the fetch_initial_data function
     mocked_fetch_initial_data = mocker.patch(
-        "afang.backtest_data_collector.fetch_initial_data"
+        "afang.database.backtest_data_collector.fetch_initial_data"
     )
     mocked_fetch_initial_data.return_value = (50, 60)
 
     # mock the fetch_most_recent_data function
     mocked_fetch_most_recent_data = mocker.patch(
-        "afang.backtest_data_collector.fetch_most_recent_data"
+        "afang.database.backtest_data_collector.fetch_most_recent_data"
     )
     mocked_fetch_most_recent_data.return_value = 70
 
     # mock the fetch_older_data function
     mocked_fetch_older_data = mocker.patch(
-        "afang.backtest_data_collector.fetch_older_data"
+        "afang.database.backtest_data_collector.fetch_older_data"
     )
     mocked_fetch_older_data.return_value = 40
 
     # mock the is_dataset_valid function
     mocked_is_dataset_valid = mocker.patch(
-        "afang.backtest_data_collector.OHLCVDatabase.is_dataset_valid"
+        "afang.database.backtest_data_collector.OHLCVDatabase.is_dataset_valid"
     )
     mocked_is_dataset_valid.return_value = True
